requirements.txt
asgiref==3.5.2
Django==4.0.6
djangorestframework==3.13.1
psycopg2==2.9.3
pyaes==1.6.1
pyasn1==0.4.8
pytz==2022.1
rsa==4.9
sqlparse==0.4.2
Telethon==1.24.0
tzdata==2022.1
